+++ client/src/Forms/ProfileForm.tsx
@@ -14,8 +14,8 @@ import { Formik, FormikProps, FormikValues } from 'formik';
 import { useSnackbar } from 'notistack';
 import React, { ReactElement, useEffect } from 'react';
 import * as Yup from 'yup';
-import { GenericError, NetworkError } from '../Components/ErrorSnackbars';
-import { FormLoading } from '../Components/Skeletons/FormLoading';
+import { GenericError, NetworkError } from '../components/ErrorSnackbars';
+import { FormLoading } from '../components/Skeletons/FormLoading';
 import { useUpdateMeMutation, useUserLazyQuery } from '../gql/queries';
 
 const useStyles = makeStyles((theme: Theme) => ({
