+++ client/src/components/Dashboard/RatingChart.tsx
