+++ client/src/components/Dashboard/ResolutionChart.tsx
