+++ client/src/components/Dashboard/TopTenSeriesCard.tsx
