+++ client/src/components/DeleteConfirmDialog.tsx
