+++ client/src/components/FileInfo.tsx
