+++ client/src/components/SeriesAutocomplete.tsx
