+++ client/src/components/SeriesInfo.tsx
