+++ client/src/components/SeriesRelatedDisplay.tsx
