+++ client/src/components/Skeletons/RatingSkeleton.tsx
