+++ client/src/components/Skeletons/SeriesInfoSkeleton.tsx
