+++ client/src/components/Skeletons/UserProgressInfoSkeleton.tsx
